Simple script for mounting with sshfs
So this is a proof of concept script I wrote to easily be able to mount
external partitions through sshfs.

Why? Because updating a milion aliases on various computers is a pain,
much easier to just update a single config file once in a while.

To anyone that feels it will benefit them, you are welcome :-)